implement redirect to overviews page #34
Instead of navigating to the index page for all "Documentation"
sections, we now add the "Overview" to the main menu (but call the
link "Documentation"). This way, the user is directly routed to the
Overview page instead of the index page.

TODO: Instruct the webserver to execute a redirect if a user
wants to navigate to /docs (navigate to /docs/overview instead).