add link to ways to start localstack container for s3 guide
While getting started with the S3 guide (and localstack) for the first time I got an error message because I didn't know that we had to run `localstack start` to start the container first. This PR adds a link to the ways to start the container. I would also recommend calling this out more explicitly for the first time users maybe as a separate step also.