Specify Docker Desktop in development-environment-setup.md
The list of required tools mentions `docker`, but that is a umbrella-term for a lot of things. Renaming it to `Docker Desktop` makes it more obvious that this specific Docker tool is required.

The additional Tip is to help people if they happen to have only the Docker engine installed. 
One might think that the Docker Engine is enough, but newer versions of the engine do not include `buildx` by default, and building the LS Docker image fails in an odd way of `buildx` cannot be found.